% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarizeddocx-package.R
\name{demo_document}
\alias{demo_document}
\title{Render and open the demo RMarkdown file}
\usage{
demo_document()
}
\value{
This function renders the built-in RMarkdown file and opens it. It
returns the path to the rendered file invisibly.
}
\description{
Render and open the demo RMarkdown file
}
