% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarizeddocx-package.R
\name{document}
\alias{document}
\title{Custom RMarkdown Docx Format}
\usage{
document(
  ...,
  reference_docx = "solarized",
  highlight = "solarized",
  pandoc_args = character(0)
)
}
\description{
Custom RMarkdown Docx Format
}
