% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarizeddocx-package.R
\name{fileresources}
\alias{fileresources}
\alias{file_reference_docx}
\alias{file_syntax_definition}
\alias{file_solarized_light_theme}
\alias{file_syntax_test}
\title{Retrieve file paths to package resources}
\usage{
file_reference_docx()

file_syntax_definition()

file_solarized_light_theme()

file_syntax_test()
}
\value{
\code{file_reference_docx()} returns the path to the reference docx file
used by this package. \code{file_syntax_definition()} returns the path to the
XML file specifying how pandoc should highlight the code in the document.
\code{file_solarized_light_theme()} returns the path to the .theme file used by
this package. \code{file_syntax_test()} returns the path to the RMarkdown file
used by \code{\link[=demo_document]{demo_document()}}.
}
\description{
Retrieve file paths to package resources
}
