% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarizeddocx-package.R
\docType{package}
\name{solarizeddocx-package}
\alias{solarizeddocx}
\alias{solarizeddocx-package}
\title{solarizeddocx: A 'Solarized' Theme for 'RMarkdown` `Docx` Documents}
\description{
An 'RMarkdown' output format for 'Microsoft' Word '.docx' documents created by 'pandoc'.
}
\author{
\strong{Maintainer}: Tristan Mahr \email{tristan.mahr@wisc.edu} (\href{https://orcid.org/0000-0002-8890-5116}{ORCID})

}
\keyword{internal}
